chore(gsdk): introduce module storage
