+++ gsdk/src/signer/calls.rs
@@ -23,37 +23,21 @@ use crate::{
     metadata::{
         calls::{BalancesCall, GearCall, SudoCall, UtilityCall},
         gear_runtime::RuntimeCall,
-        runtime_types::{
-            frame_system::pallet::Call,
-            gear_common::{ActiveProgram, Program},
-            gear_core::code::InstrumentedCode,
-            pallet_gear_bank::pallet::BankAccount,
-            sp_weights::weight_v2::Weight,
-        },
-        storage::{GearBankStorage, GearGasStorage, GearProgramStorage},
+        runtime_types::sp_weights::weight_v2::Weight,
         sudo::Event as SudoEvent,
         Event,
     },
     signer::SignerRpc,
-    utils::storage_address_bytes,
-    Api, BlockNumber, Error, GearGasNode, GearGasNodeId, GearPages, Result, TxInBlock, TxStatus,
+    Error, Result, TxInBlock, TxStatus,
 };
 use anyhow::anyhow;
-use gear_core::{
-    ids::*,
-    memory::{PageBuf, PageBufInner},
-};
-use hex::ToHex;
-use parity_scale_codec::Encode;
+use gear_core::ids::*;
 use sp_runtime::AccountId32;
 use std::sync::Arc;
 use subxt::{
     blocks::ExtrinsicEvents,
     dynamic::Value,
-    metadata::EncodeWithMetadata,
-    storage::StorageAddress,
     tx::{DynamicPayload, TxProgress},
-    utils::Static,
     Error as SubxtError, OnlineClient,
 };
 
@@ -63,10 +47,6 @@ type EventsResult = Result<ExtrinsicEvents<GearConfig>, Error>;
 /// Implementation of calls to programs/other users for [`Signer`].
 #[derive(Clone)]
 pub struct SignerCalls(pub(crate) Arc<SignerInner>);
-
-/// Implementation of storage calls for [`Signer`].
-#[derive(Clone)]
-pub struct SignerStorage(pub(crate) Arc<SignerInner>);
 
 // pallet-balances
 impl SignerCalls {
@@ -243,125 +223,6 @@ impl SignerInner {
     pub async fn sudo(&self, call: RuntimeCall) -> EventsResult {
         self.sudo_run_tx(SudoCall::Sudo, vec![Value::from(call)])
             .await
-    }
-}
-
-// pallet-system
-impl SignerStorage {
-    /// Sets storage values via calling sudo pallet
-    pub async fn set_storage(&self, items: &[(impl StorageAddress, impl Encode)]) -> EventsResult {
-        let metadata = self.0.api().metadata();
-        let mut items_to_set = Vec::with_capacity(items.len());
-        for item in items {
-            let item_key = storage_address_bytes(&item.0, &metadata)?;
-            let mut item_value_bytes = Vec::new();
-            let item_value_type_id = crate::storage::storage_type_id(&metadata, &item.0)?;
-            Static(&item.1).encode_with_metadata(
-                item_value_type_id,
-                &metadata,
-                &mut item_value_bytes,
-            )?;
-            items_to_set.push((item_key, item_value_bytes));
-        }
-
-        self.0
-            .sudo(RuntimeCall::System(Call::set_storage {
-                items: items_to_set,
-            }))
-            .await
-    }
-}
-
-// pallet-gas
-impl SignerStorage {
-    /// Writes gas total issuance into storage.
-    pub async fn set_total_issuance(&self, value: u64) -> EventsResult {
-        let addr = Api::storage_root(GearGasStorage::TotalIssuance);
-        self.set_storage(&[(addr, value)]).await
-    }
-
-    /// Writes Gear gas nodes into storage at their ids.
-    pub async fn set_gas_nodes(
-        &self,
-        gas_nodes: &impl AsRef<[(GearGasNodeId, GearGasNode)]>,
-    ) -> EventsResult {
-        let gas_nodes = gas_nodes.as_ref();
-        let mut gas_nodes_to_set = Vec::with_capacity(gas_nodes.len());
-        for gas_node in gas_nodes {
-            let addr = Api::storage(GearGasStorage::GasNodes, vec![Static(gas_node.0)]);
-            gas_nodes_to_set.push((addr, &gas_node.1));
-        }
-        self.set_storage(&gas_nodes_to_set).await
-    }
-}
-
-// pallet-gear-bank
-impl SignerStorage {
-    /// Writes given BankAccount info into storage at `AccountId32`.
-    pub async fn set_bank_account_storage(
-        &self,
-        dest: impl Into<AccountId32>,
-        value: BankAccount<u128>,
-    ) -> EventsResult {
-        let addr = Api::storage(GearBankStorage::Bank, vec![Value::from_bytes(dest.into())]);
-        self.set_storage(&[(addr, value)]).await
-    }
-}
-
-// pallet-gear-program
-impl SignerStorage {
-    /// Writes `InstrumentedCode` length into storage at `CodeId`
-    pub async fn set_code_len_storage(&self, code_id: CodeId, code_len: u32) -> EventsResult {
-        let addr = Api::storage(
-            GearProgramStorage::CodeLenStorage,
-            vec![Value::from_bytes(code_id)],
-        );
-        self.set_storage(&[(addr, code_len)]).await
-    }
-
-    /// Writes `InstrumentedCode` into storage at `CodeId`
-    pub async fn set_code_storage(&self, code_id: CodeId, code: &InstrumentedCode) -> EventsResult {
-        let addr = Api::storage(
-            GearProgramStorage::CodeStorage,
-            vec![Value::from_bytes(code_id)],
-        );
-        self.set_storage(&[(addr, code)]).await
-    }
-
-    /// Writes `GearPages` into storage at `program_id`
-    pub async fn set_gpages(
-        &self,
-        program_id: ProgramId,
-        program_pages: &GearPages,
-    ) -> EventsResult {
-        let mut program_pages_to_set = Vec::with_capacity(program_pages.len());
-        for program_page in program_pages {
-            let addr = Api::storage(
-                GearProgramStorage::MemoryPageStorage,
-                vec![
-                    subxt::dynamic::Value::from_bytes(program_id),
-                    subxt::dynamic::Value::u128(*program_page.0 as u128),
-                ],
-            );
-            let page_buf_inner = PageBufInner::try_from(program_page.1.clone())
-                .map_err(|_| Error::PageInvalid(*program_page.0, program_id.encode_hex()))?;
-            let value = PageBuf::from_inner(page_buf_inner);
-            program_pages_to_set.push((addr, value));
-        }
-        self.set_storage(&program_pages_to_set).await
-    }
-
-    /// Writes `ActiveProgram` into storage at `program_id`
-    pub async fn set_gprog(
-        &self,
-        program_id: ProgramId,
-        program: ActiveProgram<BlockNumber>,
-    ) -> EventsResult {
-        let addr = Api::storage(
-            GearProgramStorage::ProgramStorage,
-            vec![Value::from_bytes(program_id)],
-        );
-        self.set_storage(&[(addr, &Program::Active(program))]).await
     }
 }
